拆分script.js，小修index.html
