+++ app/web/static/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="zh-cn">
+<head>
+    <!-- ...existing code... -->
+</head>
+<body>
+    <!-- ...existing code... -->
+    <script src="main.js"></script>
+</body>
+</html>
