+++ app/web/static/logManager.js
@@ -1,0 +1,15 @@
+export function updateLog(logs) {
+    // ...existing code...
+}
+
+export function addLog(message, level) {
+    // ...existing code...
+}
+
+export function updateSystemLogs(logs) {
+    // ...existing code...
+}
+
+export function addSystemLogsToChat(logs) {
+    // ...existing code...
+}
