+++ app/web/templates/index.html
@@ -150,12 +150,12 @@
         </footer>
     </div>
     
-    <div class="system-logs-panel" id="systemLogsPanel">
+    <!-- <div class="system-logs-panel" id="systemLogsPanel">
         <h3>系统日志</h3>
         <div class="system-logs-container" id="systemLogsContainer">
             <p class="log-line">等待日志加载...</p>
         </div>
-    </div>
+    </div> -->
 
     <script src="{{ url_for('static', path='/script.js') }}"></script>
     <script>
